Site done with memory context

+++ Data/Dto/NutrientDto.cs
@@ -6,6 +6,6 @@ namespace Data.Dto
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double MaxIntake { get; set; }
+        public decimal MaxIntake { get; set; }
     }
 }
