+++ Fit/Models/Nutrient.cs
@@ -6,6 +6,6 @@ namespace Fit.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double MaxIntake { get; set; }
+        public decimal MaxIntake { get; set; }
     }
 }
