+++ Fit/ViewModels/Nutrient/NutrientAddViewModel.cs
@@ -3,6 +3,6 @@ namespace Fit.ViewModels.Nutrient
     public class NutrientAddViewModel
     {
         public string Name { get; set; }
-        public double MaxIntake { get; set; }
+        public decimal MaxIntake { get; set; }
     }
 }
